echo_bot/requirements.txt
python-telegram-bot==12.7.0